requirements.txt
numpy==1.20.3
scipy==1.7.1
pillow==8.3.1
numba==0.53.1
opencv-python==4.1.2.30
psutil==5.8.0
vtk==8.1.2
mayavi==4.7.2
pyqt5==5.15.6
